better changes money and plugin

+++ simulateur/simulation/models/portfolio.py
@@ -3,7 +3,7 @@ from django.db import models
 class Portfolio(models.Model):
     owner = models.OneToOneField('UserProfile', on_delete=models.CASCADE, null=True, blank=True, related_name='portfolio')
     teams = models.ManyToManyField('Team', related_name='portfolios', blank=True)
-    balance = models.FloatField(default=0.0)
+    balance = models.DecimalField(max_digits=10, decimal_places=2, default=0.00)
     stocks = models.ManyToManyField('Stock', through='StockPortfolio')
     transactions = models.ManyToManyField('TransactionHistory', related_name='portfolios', blank=True)
